+++ ProjectFreeGenre/Assets/Script/Manager/UIManager.cs
@@ -5,8 +5,8 @@ public class UIManager : Singleton<UIManager>
     public Image hpBar;
     public ItemSlotListUI itemSlotListUI;
 
-    public void UpdateHpBar(int currentHp, int maxHp) 
+    public void UpdateHpBar(float currentHp, float maxHp) 
     {
-        hpBar.fillAmount = (float)currentHp / maxHp; 
+        hpBar.fillAmount = currentHp / maxHp; 
     }
 }
